pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "renewable-fuel-twin"
version = "0.1.0"
description = "Specification-driven renewable shipping fuel market twin"
requires-python = ">=3.12"
authors = [{name = "Miro Ship"}]
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "networkx>=3.3",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
